frontend/src/app/components/login/login.component.ts
import {Component, ElementRef, OnInit, ViewChild} from '@angular/core';
import {FormBuilder, FormGroup, Validators} from "@angular/forms";
import {AuthenticationService} from "../../_service/authentication.service";
import {Router} from "@angular/router";

@Component({
  selector: 'app-login',
  templateUrl: './login.component.html',
  styleUrls: ['./login.component.scss']
})
export class LoginComponent implements OnInit {
  items;
  loginForm;

  constructor(
    private authenticationService: AuthenticationService,
    private formBuilder: FormBuilder,
    private router: Router
  ) {
    this.loginForm = this.formBuilder.group({
      username: '',
      password: ''
    });
  }

  ngOnInit() {
    if (sessionStorage.getItem('username') !== null) this.router.navigate(['/home']);
  }

  onSubmit() {
    this.authenticationService.authenticate(this.loginForm.get('username').value,this.loginForm.get('password').value)
  }
}
